+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - add ignore resource version in save dialog for conflicted resources
+- show theme loading / deserializing errors in the UI
 
 ### Bug fixes